feat(ci): tried to integrate the ci dashboard script into all testing workflows (#26694)
Would keep an eye on the CI for the next few hours if there's any issues
