--- .github/workflows/manual-tests-old.yml
@@ -1,100 +1,0 @@
-name: manual-tests-old
-on:
-  workflow_dispatch:
-    inputs:
-      test:
-        description: 'Test:'
-        default: 'ivy_tests/test_ivy/test_frontends/test_torch/test_tensor.py::test_torch_instance_arctan_,tensorflow'
-        required: true
-
-      version:
-        description: 'Version Based Testing :'
-        default: false
-        required: false
-
-      gpu:
-        description: 'Gpu based Testing :'
-        default: false
-        required: false
-
-permissions:
-  actions: read
-
-jobs:
-  run_tests_gpu:
-    if: ${{ github.event.inputs.gpu == 'true' }}
-    runs-on: self-hosted
-    steps:
-      - name: Clean repository
-        run:
-          sudo rm -fr $GITHUB_WORKSPACE && mkdir $GITHUB_WORKSPACE
-
-      - name: Checkout Ivy 🛎
-        uses: actions/checkout@v2
-        with:
-          path: ivy
-          persist-credentials: false
-          submodules: "recursive"
-          set-safe-directory: false
-
-      - name: Install Ivy
-        run: |
-          cd ivy
-          pip install .
-
-      - name: Get Job URL
-        uses: Tiryoh/gha-jobid-action@v0
-        id: jobs
-        with:
-          github_token: ${{ secrets.GITHUB_TOKEN }}
-          job_name: ${{ github.job }}
-
-      - name: Run Tests
-        id: tests
-        run: |
-          pip3 install pymongo
-          cd ivy
-          mkdir .ivy
-          touch .ivy/key.pem
-          echo -n ${{ secrets.USER_API_KEY }} > .ivy/key.pem
-          python3 setup_tests.py ${{ github.event.inputs.test }}
-          python3 run_tests.py ${{ secrets.REDIS_CONNECTION_URL }} ${{ secrets.REDIS_PASSWORD }} ${{ secrets.MONGODB_PASSWORD }} 'false' ${{ github.event.inputs.gpu }} ${{ github.run_id }} 'false' ${{ steps.jobs.outputs.html_url }}
-        continue-on-error: true
-
-      - name: Check on failures
-        if: steps.tests.outcome != 'success'
-        run: exit 1
-
-  run_tests_cpu:
-    if: ${{ github.event.inputs.gpu == 'false' }}
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout Ivy 🛎
-        uses: actions/checkout@v2
-        with:
-          path: ivy
-          persist-credentials: false
-          submodules: "recursive"
-
-      - name: Get Job URL
-        uses: Tiryoh/gha-jobid-action@v0
-        id: jobs
-        with:
-          github_token: ${{ secrets.GITHUB_TOKEN }}
-          job_name: ${{ github.job }}
-
-      - name: Run Tests
-        id: tests
-        run: |
-          pip3 install pymongo
-          cd ivy
-          mkdir .ivy
-          touch .ivy/key.pem
-          echo -n ${{ secrets.USER_API_KEY }} > .ivy/key.pem
-          python setup_tests.py "${{ github.event.inputs.test }}"
-          python run_tests.py ${{ secrets.REDIS_CONNECTION_URL }} ${{ secrets.REDIS_PASSWORD }} ${{ secrets.MONGODB_PASSWORD }} ${{ github.event.inputs.version}} 'false' ${{ github.run_id }} 'false' ${{ steps.jobs.outputs.html_url }}
-        continue-on-error: true
-
-      - name: Check on failures
-        if: steps.tests.outcome != 'success'
-        run: exit 1
